+++ staging/vhost-device-gpu/src/main.rs
@@ -4,34 +4,53 @@
 //
 // SPDX-License-Identifier: Apache-2.0 or BSD-3-Clause
 
-// Rust vmm container (https://github.com/rust-vmm/rust-vmm-container) doesn't
-// have tools to do a musl build at the moment, and adding that support is
-// tricky as well to the container. Skip musl builds until the time pre-built
-// rutabaga library is available for musl.
+// See the cfg(target_env = "gnu") explanation in lib.rs
 #[cfg(target_env = "gnu")]
 pub mod gnu_main {
     use std::{path::PathBuf, process::exit};
 
-    use clap::Parser;
-    use log::{error, info};
-    use thiserror::Error as ThisError;
+    use clap::{ArgAction, Parser, ValueEnum};
+    use log::error;
     use vhost_device_gpu::{
-        device::{self, VhostUserGpuBackend},
-        GpuConfig, GpuMode,
+        start_backend, GpuCapset, GpuConfig, GpuConfigError, GpuFlags, GpuMode,
     };
-    use vhost_user_backend::VhostUserDaemon;
-    use vm_memory::{GuestMemoryAtomic, GuestMemoryMmap};
 
-    type Result<T> = std::result::Result<T, Error>;
+    #[derive(ValueEnum, Debug, Copy, Clone, Eq, PartialEq)]
+    #[repr(u64)]
+    pub enum CapsetName {
+        /// [virglrenderer] OpenGL implementation, superseded by Virgl2
+        Virgl = GpuCapset::VIRGL.bits(),
 
-    #[derive(Debug, ThisError)]
-    pub enum Error {
-        #[error("Could not create backend: {0}")]
-        CouldNotCreateBackend(device::Error),
-        #[error("Could not create daemon: {0}")]
-        CouldNotCreateDaemon(vhost_user_backend::Error),
-        #[error("Fatal error: {0}")]
-        ServeFailed(vhost_user_backend::Error),
+        /// [virglrenderer] OpenGL implementation
+        Virgl2 = GpuCapset::VIRGL2.bits(),
+
+        /// [gfxstream] Vulkan implementation (partial support only){n}
+        /// NOTE: Can only be used for 2D display output for now, there is no
+        /// hardware acceleration yet
+        #[cfg(feature = "gfxstream")]
+        GfxstreamVulkan = GpuCapset::GFXSTREAM_VULKAN.bits(),
+
+        /// [gfxstream] OpenGL ES implementation (partial support only){n}
+        /// NOTE: Can only be used for 2D display output for now, there is no
+        /// hardware acceleration yet
+        #[cfg(feature = "gfxstream")]
+        GfxstreamGles = GpuCapset::GFXSTREAM_GLES.bits(),
+    }
+
+    impl From<CapsetName> for GpuCapset {
+        fn from(capset_name: CapsetName) -> GpuCapset {
+            GpuCapset::from_bits(capset_name as u64)
+                .expect("Internal error: CapsetName enum is incorrectly defined")
+        }
+    }
+
+    pub fn capset_names_into_capset(
+        capset_names: impl IntoIterator<Item = CapsetName>,
+    ) -> GpuCapset {
+        capset_names
+            .into_iter()
+            .map(CapsetName::into)
+            .fold(GpuCapset::empty(), GpuCapset::union)
     }
 
     #[derive(Parser, Debug)]
@@ -40,41 +59,87 @@ pub mod gnu_main {
         /// vhost-user Unix domain socket.
         #[clap(short, long, value_name = "SOCKET")]
         pub socket_path: PathBuf,
+
+        /// The mode specifies which backend implementation to use
         #[clap(short, long, value_enum)]
         pub gpu_mode: GpuMode,
+
+        /// Comma separated list of enabled capsets
+        #[clap(short, long, value_delimiter = ',')]
+        pub capset: Option<Vec<CapsetName>>,
+
+        #[clap(flatten)]
+        pub flags: GpuFlagsArgs,
     }
 
-    impl From<GpuArgs> for GpuConfig {
-        fn from(args: GpuArgs) -> Self {
-            let socket_path = args.socket_path;
-            let gpu_mode: GpuMode = args.gpu_mode;
+    #[derive(Parser, Debug)]
+    #[allow(clippy::struct_excessive_bools)]
+    pub struct GpuFlagsArgs {
+        /// Enable backend to use EGL
+        #[clap(
+            long,
+            action=ArgAction::Set,
+            default_value_t = GpuFlags::new_default().use_egl
+        )]
+        pub use_egl: bool,
 
-            GpuConfig::new(socket_path, gpu_mode)
+        /// Enable backend to use GLX
+        #[clap(
+            long,
+            action=ArgAction::Set,
+            default_value_t = GpuFlags::new_default().use_glx
+        )]
+        pub use_glx: bool,
+
+        /// Enable backend to use GLES
+        #[clap(
+            long,
+            action=ArgAction::Set,
+            default_value_t = GpuFlags::new_default().use_gles
+        )]
+        pub use_gles: bool,
+
+        /// Enable surfaceless backend option
+        #[clap(
+            long,
+            action = ArgAction::Set,
+            default_value_t = GpuFlags::new_default().use_surfaceless
+        )]
+        pub use_surfaceless: bool,
+    }
+
+    impl From<GpuFlagsArgs> for GpuFlags {
+        fn from(args: GpuFlagsArgs) -> Self {
+            GpuFlags {
+                use_egl: args.use_egl,
+                use_glx: args.use_glx,
+                use_gles: args.use_gles,
+                use_surfaceless: args.use_surfaceless,
+            }
         }
     }
 
-    pub fn start_backend(config: &GpuConfig) -> Result<()> {
-        info!("Starting backend");
-        let socket = config.socket_path();
-        let backend = VhostUserGpuBackend::new(config).map_err(Error::CouldNotCreateBackend)?;
-
-        let mut daemon = VhostUserDaemon::new(
-            "vhost-device-gpu-backend".to_string(),
-            backend.clone(),
-            GuestMemoryAtomic::new(GuestMemoryMmap::new()),
-        )
-        .map_err(Error::CouldNotCreateDaemon)?;
-
-        backend.set_epoll_handler(&daemon.get_epoll_handlers());
-
-        daemon.serve(socket).map_err(Error::ServeFailed)?;
-        Ok(())
+    pub fn config_from_args(args: GpuArgs) -> Result<(PathBuf, GpuConfig), GpuConfigError> {
+        let flags = GpuFlags::from(args.flags);
+        let capset = args.capset.map(capset_names_into_capset);
+        let config = GpuConfig::new(args.gpu_mode, capset, flags)?;
+        Ok((args.socket_path, config))
     }
 
     pub fn main() {
         env_logger::init();
 
-        if let Err(e) = start_backend(&GpuConfig::from(GpuArgs::parse())) {
+        let args = GpuArgs::parse();
+
+        let (socket_path, config) = match config_from_args(args) {
+            Ok(config) => config,
+            Err(e) => {
+                error!("{e}");
+                exit(1);
+            }
+        };
+
+        if let Err(e) = start_backend(&socket_path, config) {
             error!("{e}");
             exit(1);
         }
@@ -94,39 +159,65 @@ fn main() {}
 mod tests {
     use std::path::Path;
 
-    use assert_matches::assert_matches;
-    use tempfile::tempdir;
-    use vhost_device_gpu::{GpuConfig, GpuMode};
+    use clap::{Parser, ValueEnum};
+    use vhost_device_gpu::{GpuCapset, GpuFlags, GpuMode};
 
-    use super::gnu_main::{start_backend, Error, GpuArgs};
+    use super::gnu_main::*;
 
-    impl GpuArgs {
-        pub(crate) fn from_args(path: &Path) -> Self {
-            Self {
-                socket_path: path.to_path_buf(),
-                gpu_mode: GpuMode::Gfxstream,
-            }
+    #[test]
+    fn test_capset_enum_in_sync_with_capset_bitset() {
+        // Convert each GpuCapset into CapsetName
+        for capset in GpuCapset::all().iter() {
+            let display_name = capset.to_string();
+            let capset_name = CapsetName::from_str(&display_name, false).unwrap();
+            let resulting_capset: GpuCapset = capset_name.into();
+            assert_eq!(resulting_capset, capset);
+        }
+
+        // Convert each CapsetName into GpuCapset
+        for capset_name in CapsetName::value_variants().iter().cloned() {
+            let resulting_capset: GpuCapset = capset_name.into(); // Would panic! if the definition is incorrect
+            assert_eq!(resulting_capset.bits(), capset_name as u64)
         }
     }
 
     #[test]
-    fn test_parse_successful() {
-        let test_dir = tempdir().expect("Could not create a temp test directory.");
-        let socket_path = test_dir.path().join("vgpu.sock");
-
-        let cmd_args = GpuArgs::from_args(socket_path.as_path());
-        let config = GpuConfig::from(cmd_args);
-
-        assert_eq!(config.socket_path(), socket_path);
+    fn test_default_cli_flags() {
+        // The default CLI flags should match GpuFlags::default()
+        let args: &[&str] = &[];
+        let flag_args = GpuFlagsArgs::parse_from(args);
+        let flags: GpuFlags = flag_args.into();
+        assert_eq!(flags, GpuFlags::default());
     }
 
     #[test]
-    fn test_fail_listener() {
-        // This will fail the listeners and thread will panic.
-        let socket_name = Path::new("/proc/-1/nonexistent");
-        let cmd_args = GpuArgs::from_args(socket_name);
-        let config = GpuConfig::from(cmd_args);
+    fn test_config_from_args() {
+        let expected_path = Path::new("/some/test/path");
+        let args = GpuArgs {
+            socket_path: expected_path.into(),
+            gpu_mode: GpuMode::VirglRenderer,
+            capset: Some(vec![CapsetName::Virgl, CapsetName::Virgl2]),
+            flags: GpuFlagsArgs {
+                use_egl: false,
+                use_glx: true,
+                use_gles: false,
+                use_surfaceless: false,
+            },
+        };
 
-        assert_matches!(start_backend(&config).unwrap_err(), Error::ServeFailed(_));
+        let (socket_path, config) = config_from_args(args).unwrap();
+
+        assert_eq!(socket_path, expected_path);
+        assert_eq!(
+            *config.flags(),
+            GpuFlags {
+                use_egl: false,
+                use_glx: true,
+                use_gles: false,
+                use_surfaceless: false,
+            }
+        );
+        assert_eq!(config.gpu_mode(), GpuMode::VirglRenderer);
+        assert_eq!(config.capsets(), GpuCapset::VIRGL | GpuCapset::VIRGL2)
     }
 }