+++ staging/vhost-device-gpu/CHANGELOG.md
@@ -2,8 +2,11 @@
 ## Unreleased
 
 ### Added
+- Add command line arguments for specifying the exact capsets to enable and flags to pass to the backend. 
 
 ### Changed
+- Rename --gpu-mode virgl-renderer to "virglrenderer"
+- Do not implicitly enable all capsets in Rutabaga, only enable capsets that are at least partially supported.
 
 ### Fixed